# Conflicts:
#	activity/pom.xml
#	activity/src/main/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityOffer.java
#	activity/src/main/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityProvider.java
#	activity/src/main/java/pt/ulisboa/tecnico/softeng/activity/domain/Booking.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityConstructorMethodTest.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityMatchAgeMethodTest.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityOfferConstructorMethodTest.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityOfferGetBookingMethodTest.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityOfferHasVacancyMethodTest.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityOfferMatchDateMethodTest.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityProviderActivityReservationDataMethodTest.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityProviderCancelReservationMethodTest.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityProviderConstructorMethodTest.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/ActivityProviderFindOfferMethodTest.java
#	activity/src/test/java/pt/ulisboa/tecnico/softeng/activity/domain/BookingContructorMethodTest.java
#	bank/pom.xml
#	bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/dataobjects/BankOperationData.java
#	bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/domain/Account.java
#	bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/domain/Bank.java
#	bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/domain/Operation.java
#	bank/src/test/java/pt/ulisboa/tecnico/softeng/bank/domain/AccountContructorMethodTest.java
#	bank/src/test/java/pt/ulisboa/tecnico/softeng/bank/domain/BankGetAccountMethodTest.java
#	bank/src/test/java/pt/ulisboa/tecnico/softeng/bank/domain/BankGetOperationDataMethodTest.java
#	broker/pom.xml
#	broker/src/main/dml/broker.dml
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/Adventure.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/BookRoomState.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/Broker.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/BulkRoomBooking.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/CancelledState.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/ConfirmedState.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/ProcessPaymentState.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/ReserveActivityState.java
#	broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/domain/UndoState.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/AdventureConstructorMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/AdventureProcessMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/AdventureSequenceTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/BookRoomStateMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/BrokerConstructorMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/BulkRoomBookingGetReferenceMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/BulkRoomBookingProcessBookingMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/CancelledStateProcessMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/ConfirmedStateProcessMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/ProcessPaymentStateProcessMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/ReserveActivityStateProcessMethodTest.java
#	broker/src/test/java/pt/ulisboa/tecnico/softeng/broker/domain/UndoStateProcessMethodTest.java
#	hotel/pom.xml
#	hotel/src/main/java/pt/ulisboa/tecnico/softeng/hotel/domain/Booking.java
#	hotel/src/main/java/pt/ulisboa/tecnico/softeng/hotel/domain/Hotel.java
#	hotel/src/main/java/pt/ulisboa/tecnico/softeng/hotel/domain/Room.java
#	hotel/src/test/java/pt/ulisboa/tecnico/softeng/hotel/domain/BookingConflictMethodTest.java
#	hotel/src/test/java/pt/ulisboa/tecnico/softeng/hotel/domain/BookingConstructorTest.java
#	hotel/src/test/java/pt/ulisboa/tecnico/softeng/hotel/domain/HotelBulkBookingMethodTest.java
#	hotel/src/test/java/pt/ulisboa/tecnico/softeng/hotel/domain/HotelCancelBookingMethodTest.java
#	hotel/src/test/java/pt/ulisboa/tecnico/softeng/hotel/domain/HotelConstructorTest.java
#	hotel/src/test/java/pt/ulisboa/tecnico/softeng/hotel/domain/HotelGetRoomBookingDataMethodTest.java
#	hotel/src/test/java/pt/ulisboa/tecnico/softeng/hotel/domain/HotelHasVacancyMethodTest.java
#	hotel/src/test/java/pt/ulisboa/tecnico/softeng/hotel/domain/RoomConstructorMethodTest.java
#	hotel/src/test/java/pt/ulisboa/tecnico/softeng/hotel/domain/RoomGetBookingMethodTest.java
#	hotel/src/test/java/pt/ulisboa/tecnico/softeng/hotel/domain/RoomReserveMethodTest.java
#	pom.xml

+++ bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/dataobjects/BankOperationData.java
@@ -8,7 +8,7 @@ public class BankOperationData {
 	private String reference;
 	private String type;
 	private String iban;
-	private int value;
+	private double value;
 	private DateTime time;
 
 	public BankOperationData() {
@@ -46,7 +46,7 @@ public class BankOperationData {
 		this.iban = iban;
 	}
 
-	public int getValue() {
+	public double getValue() {
 		return this.value;
 	}
