+++ bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/domain/Bank.java
@@ -92,7 +92,7 @@ public class Bank extends Bank_Base {
 		return null;
 	}
 
-	public static String processPayment(String IBAN, int amount) {
+	public static String processPayment(String IBAN, double amount) {
 		for (Bank bank : FenixFramework.getDomainRoot().getBankSet()) {
 			if (bank.getAccount(IBAN) != null) {
 				return bank.getAccount(IBAN).withdraw(amount);
